Added v0.1.1 Pub
All source files available, exe is just a compiled version through pyinstaller. It makes it easier for usage when tracking inventory. Inventory.csv is included to show formatting and program.log as well to show some of its features.

I will be improving this program, allowing it to be run remotely on mysql or locally csv, and this project will be maintained as it is for a personal program.